Move recurring action related code to its own controller
